separated spec files for column
