+++ cypress/integration/v2/blocks/column/other-tests.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { switchColumnLayout } from '.'
+
+describe( 'Column Block ( Other Tests )', registerTests( [ switchColumnLayout ] ) )
